third_party/WebKit/LayoutTests/fast/shapes/parsing/parsing-shape-outside-none.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#target { shape-outside: url(//fail); }
#target { shape-outside: none; }
</style>
<div id="target"></div>
<script>
test(function() {
  assert_equals(getComputedStyle(target).shapeOutside, 'none');
}, 'Test setting the none value for shape-outside');
</script>
